Update dspprofiles.md
When following the original instructions here I was unable to install a profile that was generated because I had sequences already in the sequence window when I added new sequences.  So I included instructions here for how to make sure the sequence window is also clear before adding new sequences.